#16 - Add build config * Move config variables, banners and supported browsers objects to build.conf from gulpfile. * Add build.conf,js.
